backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
httpx
pytest
python-dotenv==1.0.1
